+++ fastcrypto-vdf/benches/class_group.rs
@@ -4,7 +4,8 @@
 use criterion::measurement::Measurement;
 use criterion::{criterion_group, criterion_main, BatchSize, BenchmarkGroup, Criterion};
 use fastcrypto::groups::Doubling;
-use fastcrypto_vdf::class_group::{Discriminant, QuadraticForm};
+use fastcrypto_vdf::class_group::discriminant::Discriminant;
+use fastcrypto_vdf::class_group::QuadraticForm;
 use fastcrypto_vdf::hash_prime::{DefaultPrimalityCheck, PrimalityCheck};
 use fastcrypto_vdf::ParameterizedGroupElement;
 use num_bigint::{BigInt, UniformBigInt};
