+++ fastcrypto-vdf/benches/vdf.rs
@@ -7,10 +7,11 @@ extern crate criterion;
 use criterion::measurement::Measurement;
 use criterion::{BenchmarkGroup, BenchmarkId, Criterion};
 use fastcrypto::groups::multiplier::windowed::WindowedScalarMultiplier;
-use fastcrypto_vdf::class_group::{Discriminant, QuadraticForm};
+use fastcrypto_vdf::class_group::discriminant::Discriminant;
+use fastcrypto_vdf::class_group::QuadraticForm;
 use fastcrypto_vdf::hash_prime::DefaultPrimalityCheck;
-use fastcrypto_vdf::vdf::wesolowski::CHALLENGE_SIZE;
-use fastcrypto_vdf::vdf::wesolowski::{FastVerifier, StrongFiatShamir, StrongVDF};
+use fastcrypto_vdf::vdf::wesolowski::fiat_shamir::StrongFiatShamir;
+use fastcrypto_vdf::vdf::wesolowski::{FastVerifier, StrongVDF, CHALLENGE_SIZE};
 use fastcrypto_vdf::vdf::VDF;
 use fastcrypto_vdf::Parameter;
 use num_bigint::BigInt;
