+++ fastcrypto-vdf/src/lib.rs
@@ -12,9 +12,6 @@ pub mod class_group;
 pub mod vdf;
 
 #[cfg(any(test, feature = "experimental"))]
-pub mod hash_prime;
-
-#[cfg(any(test, feature = "experimental"))]
 pub mod math;
 
 /// This trait is implemented by types which can be used as parameters for a parameterized group.
