+++ fastcrypto-vdf/src/math/mod.rs
@@ -3,5 +3,7 @@
 
 pub mod crt;
 pub mod extended_gcd;
+#[cfg(any(test, feature = "experimental"))]
+pub mod hash_prime;
 pub mod jacobi;
 pub mod modular;
